requirements.txt
beautifulsoup4
requests
google-cloud-language
google-cloud-vision
